+++ lib/ui/screens/Home/home_screen_controller.dart
@@ -19,6 +19,7 @@ import '/services/library_sync_service.dart';
 import '/services/music_service.dart';
 import '/ui/player/player_controller.dart';
 import '/ui/screens/Search/search_screen_controller.dart';
+import '../../../utils/app_l10n.dart';
 import '../../../utils/helper.dart';
 import '../../../utils/server_storage.dart';
 import '../../../utils/update_check_flag_file.dart';
@@ -1025,7 +1026,7 @@ class HomeScreenController extends GetxController {
           '[RECOVERABLE][opId=home.startRadio.library] Failed to get library song: $e\n$st');
     }
 
-    Get.snackbar('', 'Add some favorites to start radio');
+    Get.snackbar('', l10nFromPrefs().addFavoritesToStartRadio);
   }
 
   void disposeDetachedScrollControllers({bool disposeAll = false}) {